Chore extract common findAll to lib/database
