+++ src/components/department-repo.js
@@ -1,4 +1,4 @@
-const { db, findById, store } = require('lib/database')
+const { findAll, findById, store } = require('lib/database')
 
 const table = 'departments'
 
@@ -6,13 +6,4 @@ module.exports = {
   findAll: findAll(table),
   findById: findById(table),
   store: store(table)
-}
-
-function findAll (table) {
-  return async function findAll ({ projectId }) {
-    return await db
-      .select()
-      .table(table)
-      .where('project', '=', projectId)
-  }
 }
